backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
